hrabovskyy_BOT/requirements.txt
python-telegram-bot==20.5
aiohttp~=3.12.7
apscheduler~=3.11.0
requests~=2.32.3